refactor: imporve components naming on blm
