+++ app/layout/scenarios/edit/parameters/blm-calibration/chart/component.tsx
