+++ app/layout/scenarios/edit/parameters/blm-calibration/chart/constants.ts
