+++ app/layout/scenarios/edit/parameters/blm-calibration/chart/index.ts
