+++ app/layout/scenarios/edit/parameters/blm-calibration/image-modal/component.tsx
