+++ app/layout/scenarios/edit/parameters/blm-calibration/image-modal/index.ts
