+++ app/layout/scenarios/edit/parameters/blm-calibration/settings-chart/index.ts
